Change the Way the Resulting Markdown Header is Created
Main Changes:
In addition to a "id_key" now also a "title_key" is required in the _config.yml
The "title_key" gives the name of the field within a collection entry which will
be set as the pages' title via the Markdown frontmatter "title".
The collection entry itself will no longer be pasted 1:1 into the Markdown
frontmatter but rather be a subobject of the "data" field.

Minor Changes:
	* Refactored some modifiers into statements for better reading flow.
	* Added parentheses for easier discerning of method application
	* Renamed iterators to be more meaningful
	* Renamed "overwrite_pages" to "remove_output_dir" to better match
	  its actual function
	* Added documentation comments