Generalize the Handling of Direct and Indirect Config Keys
A new "do_resolve" key has been introduced.
The key is supposed to hold a mapping from a key to a
foreign key in the processed data set. The foreign key will be
looked up in the data set and - for each data entry -  be
replaced with the value for that foreign key.

The "title_key" has been removed as its fuctionality is now
subsumed by "do_resolve".

The "frontmatter_extra" keys have been renamed into "do_copy"
to accomodate this new way of doing things.

The "id_key" remains unchanged since it has extra functionality.
The "layout" key remains unchanged since it is subject to
additional validation.